test: written unit test for DrawerStyled
